refactor: Reorganize code for creating entities
this creates a base-entity directory that holds the common factory and resolve functions for
entities
