+++ packages/data-point/lib/entity-types/entity-hash/factory.js
@@ -4,7 +4,7 @@ const _ = require('lodash')
 const createTransform = require('../../transform-expression').create
 const deepFreeze = require('deep-freeze')
 const parseCompose = require('../parse-compose')
-const helpers = require('../../helpers')
+const createBaseEntity = require('../base-entity').create
 
 /**
  * @class
@@ -101,7 +101,7 @@ function validateCompose (entityId, compose, validKeys) {
 function create (spec, id) {
   validateComposeVsInlineModifiers(spec, modifierKeys)
 
-  const entity = helpers.createEntity(EntityHash, spec, id)
+  const entity = createBaseEntity(EntityHash, spec, id)
 
   const compose = parseCompose.parse(spec, modifierKeys)
   validateCompose(entity.id, compose, modifierKeys)
