--- packages/data-point/lib/entity-types/resolve-entity/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./resolve-entity')
